+++ src/context/mode-context.tsx
@@ -8,6 +8,7 @@ import {
     ReactNode,
 } from "react";
 import { useAuth } from "./auth-context";
+import {base64ToFile, downloadFile} from "../utils/utils";
 
 type ModeContextPropsType = {
     mode: boolean;
@@ -27,6 +28,8 @@ function ModeProvider(props: ModeProviderProps) {
     const [selectedFile, setSelectedFile] = useState<File>();
     const [percentage, setPercentage] = useState<string>("0.0");
     const [appearance, setAppearance] = useState<string>("light");
+    const [documentId, setDocumentId] = useState<string>('');
+    const [slidingWidth, setSlidingWidth] = useState<number>(0)
     // @ts-ignore
     const { user } = useAuth();
 
@@ -36,8 +39,7 @@ function ModeProvider(props: ModeProviderProps) {
         // @ts-ignore
         formData.append("file", selectedFile);
         //TODO:  Upload user File
-        axios
-            .post(`http://localhost:8080/api/file?uid=${user?.uid}`, formData, {
+        axios.post(`http://localhost:8080/api/file?uid=${user?.uid}`, formData, {
                 onDownloadProgress: (progressEvent) => {
                     const logVal: string =
                         progressEvent.event.target.responseText.split("\n");
@@ -58,16 +60,36 @@ function ModeProvider(props: ModeProviderProps) {
                             }
                         }
                     }
-                    console.log(logVal[i]);
                     const percentageData = logVal[i].split(":");
+                    const percentageDatasplited =  percentageData[percentageData.length - 1].split("+");
                     const percentageNum: string = parseFloat(
-                        percentageData[percentageData.length - 1]
+                        percentageDatasplited[0]
                     ).toFixed(2);
+                    const slidingWidthOnPercentage = parseInt(String(((parseFloat(percentageNum) / 100.0) * 450)))
+                    setDocumentId(percentageDatasplited[percentageDatasplited.length - 1]);
                     setPercentage(percentageNum);
+                    setSlidingWidth(slidingWidthOnPercentage)
                 },
             })
             .then((res) => {
                 console.log(res);
+                setSlidingWidth(450);
+                setPercentage('100');
+            })
+            .catch((err) => {
+                console.log(err);
+            });
+    };
+
+    const handleDownloadFile = () => {
+        //TODO: download by file id
+        axios
+            .get(
+                `${process.env.REACT_APP_API_URL}/api/file/single-file?file_id=${documentId}`
+            )
+            .then((res) => {
+                console.log(res.data);
+                downloadFile(base64ToFile(res.data, "", ""), "test.docx");
             })
             .catch((err) => {
                 console.log(err);
@@ -84,6 +106,7 @@ function ModeProvider(props: ModeProviderProps) {
         onFileUploadHandle,
         appearance,
         setAppearance,
+        slidingWidth
     };
     return (
         <ModeContext.Provider {...props} value={value}></ModeContext.Provider>
