+++ src/pages/HomePage.tsx
@@ -1,19 +1,24 @@
-import React from "react";
+import React, {useContext, useEffect, useState} from "react";
 
 import HeaderUserInput from "../module/HeaderUserInput";
 import { Button } from "../components/button";
 import UserTextInput from "../module/UserTextInput";
 import { useMode } from "../context/mode-context";
 import UserFileInput from "../module/UserFileInput";
+import axios from "axios/index";
+import {AuthContext, AuthContextPropsType} from "../context/auth-context";
 
 const HomePage: React.FC = () => {
     // @ts-ignore
-    const { mode, onFileUploadHandle, appearance } = useMode();
+    const { mode, onFileUploadHandle, appearance, percentage, slidingWidth } = useMode();
+    // @ts-ignore
+    const { user } = useContext<AuthContextPropsType>(AuthContext);
 
     //Functions
     const hanldeTextUpload = () => {
         console.log("hello");
     };
+
 
     return (
         <div className="flex flex-col w-full h-full gap-4 px-4 py-10 mx-auto overflow-hidden">
@@ -32,6 +37,14 @@ const HomePage: React.FC = () => {
                     <UserFileInput className={`${appearance}`}></UserFileInput>
                 )}
             </div>
+            {mode ? <div className="w-[450px] h-10 mx-auto rounded-xl border-2 border-green-600">
+                <div style={{ width: `${slidingWidth}px` }} className={` h-full bg-green-400  rounded-xl transition-all duration-300`}>
+
+                </div>
+                <div className="w-[120px] mx-auto text-lg">
+                    <label className="">Loading {percentage}%</label>
+                </div>
+            </div> : null}
             <Button
                 className="mt-auto"
                 onClick={mode ? onFileUploadHandle : hanldeTextUpload}>
